pyproject.toml
[tool.poetry]
name = "ups017"
version = "0.1.0"
description = ""
authors = ["Marta Krogulec <marta.krogulec@xebia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.0"
flask = "2.2"
email-validator = "^2.0.0.post2"
factory-boy = "^3.2.1"
sqlalchemy = "^2.0.18"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
connexion = {extras = ["flask"], version = "^3.0.6"}
swagger-ui-bundle = "^1.1.0"
uvicorn = "^0.29.0"
werkzeug = "2.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pylint = "^2.17.4"
mypy = "^1.3.0"
black = "^23.3.0"
isort = "^5.12.0"
freezegun = "^1.2.2"

[tool.pylint.messages_control]
disable = ["missing-module-docstring", "missing-class-docstring", "too-few-public-methods"]

[tool.mypy]
plugins = [ "pydantic.mypy",]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pylint.MASTER]
init-hook="import os, sys, pathlib;sys.path.append(pathlib.Path(os.getcwd()) / 'src')"
extension-pkg-whitelist = "pydantic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

